source/pyproject.toml
[project]
name = "personal-accounting"
version = "0.1.0"
description = ""
authors = [
    {name = "darkimonus kimon.sum@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-yasg (>=1.21.8,<2.0.0)",
    "social-auth-app-django (>=5.4.2,<6.0.0)",
    "drf-social-oauth2 (>=3.1.0,<4.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psutil (>=6.1.1,<7.0.0)",
    "django-redis (>=5.4.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
